refactor(website): tag input component
